save upperdir

+++ snapshots/overlay/overlay.go
@@ -26,9 +26,11 @@ import (
 	"path/filepath"
 	"strings"
 	"syscall"
+	"time"
 
 	"github.com/containerd/containerd/mount"
 	"github.com/containerd/containerd/snapshots"
+	"github.com/containerd/containerd/snapshots/overlay/disk"
 	"github.com/containerd/containerd/snapshots/overlay/overlayutils"
 	"github.com/containerd/containerd/snapshots/storage"
 	"github.com/containerd/continuity/fs"
@@ -41,11 +43,29 @@ import (
 // the change set between this snapshot and its parent is stored.
 const upperdirKey = "containerd.io/snapshot/overlay.upperdir"
 
+var (
+	notebookLabelKey     = "type.system.hero.ai"
+	versionKey           = "lastversions.system.hero.ai"
+	notebookLabelValue   = "notebook"
+	notebookActionKey    = "command.system.hero.ai"
+	notebookStartValue   = "start"
+	notebookNameLabelKey = "jobid.system.hero.ai"
+	DefaultUpperdirRoot  = "/var/lib/disk"
+	DefaultDiskDb        = "disk_manager"
+	DefaultAddress       = "disk_manager.sock"
+	DefaultRootfsSize    = 30
+	DefaultTimerView     = 5 * time.Second
+)
+
 // SnapshotterConfig is used to configure the overlay snapshotter instance
 type SnapshotterConfig struct {
 	asyncRemove   bool
 	upperdirLabel bool
+	upperdirRoot  string
+	diskTimeout   int
+	rootfsQuota   int
 	mountOptions  []string
+	address       string
 }
 
 // Opt is an option to configure the overlay snapshotter
@@ -69,6 +89,37 @@ func WithUpperdirLabel(config *SnapshotterConfig) error {
 	return nil
 }
 
+func WithDiskTimeout(diskTimeout int) Opt {
+	return func(config *SnapshotterConfig) error {
+		config.diskTimeout = diskTimeout
+		return nil
+	}
+}
+
+func WithDiskAddress(address string) Opt {
+	return func(config *SnapshotterConfig) error {
+		config.address = address
+		return nil
+	}
+}
+
+func WithUpperdirRoot(ro string) Opt {
+	return func(config *SnapshotterConfig) error {
+		if err := os.MkdirAll(ro, 0700); err != nil {
+			return err
+		}
+		config.upperdirRoot = ro
+		return nil
+	}
+}
+
+func WithRootfsQuota(quotaSize int) Opt {
+	return func(config *SnapshotterConfig) error {
+		config.rootfsQuota = quotaSize
+		return nil
+	}
+}
+
 // WithMountOptions defines the default mount options used for the overlay mount.
 // NOTE: Options are not applied to bind mounts.
 func WithMountOptions(options []string) Opt {
@@ -80,10 +131,13 @@ func WithMountOptions(options []string) Opt {
 
 type snapshotter struct {
 	root          string
+	upperdirRoot  string
+	rootfsQuota   int
 	ms            *storage.MetaStore
 	asyncRemove   bool
 	upperdirLabel bool
 	options       []string
+	diskClient    *disk.DisksClient
 }
 
 // NewSnapshotter returns a Snapshotter which uses overlayfs. The overlayfs
@@ -131,12 +185,20 @@ func NewSnapshotter(root string, opts ...Opt) (snapshots.Snapshotter, error) {
 		config.mountOptions = append(config.mountOptions, "index=off")
 	}
 
+	fmt.Printf("grpc address: %s\n", config.address)
+	diskClient, err := disk.NewDisksClient(config.address)
+	if err != nil {
+		log.G(context.TODO()).Warn("failed to connect disk grpc")
+	}
 	return &snapshotter{
 		root:          root,
+		upperdirRoot:  config.upperdirRoot,
+		rootfsQuota:   config.rootfsQuota,
 		ms:            ms,
 		asyncRemove:   config.asyncRemove,
 		upperdirLabel: config.upperdirLabel,
 		options:       config.mountOptions,
+		diskClient:    diskClient,
 	}, nil
 }
 
@@ -173,7 +235,7 @@ func (o *snapshotter) Stat(ctx context.Context, key string) (snapshots.Info, err
 		if info.Labels == nil {
 			info.Labels = make(map[string]string)
 		}
-		info.Labels[upperdirKey] = o.upperPath(id)
+		info.Labels[upperdirKey] = o.upperPath(id, info.Labels)
 	}
 
 	return info, nil
@@ -199,7 +261,7 @@ func (o *snapshotter) Update(ctx context.Context, info snapshots.Info, fieldpath
 		if info.Labels == nil {
 			info.Labels = make(map[string]string)
 		}
-		info.Labels[upperdirKey] = o.upperPath(id)
+		info.Labels[upperdirKey] = o.upperPath(id, info.Labels)
 	}
 
 	if err := t.Commit(); err != nil {
@@ -228,7 +290,7 @@ func (o *snapshotter) Usage(ctx context.Context, key string) (snapshots.Usage, e
 	}
 
 	if info.Kind == snapshots.KindActive {
-		upperPath := o.upperPath(id)
+		upperPath := o.upperPath(id, info.Labels)
 		du, err := fs.DiskUsage(ctx, upperPath)
 		if err != nil {
 			// TODO(stevvooe): Consider not reporting an error in this case.
@@ -281,12 +343,12 @@ func (o *snapshotter) Commit(ctx context.Context, name, key string, opts ...snap
 	}()
 
 	// grab the existing id
-	id, _, _, err := storage.GetInfo(ctx, key)
+	id, i, _, err := storage.GetInfo(ctx, key)
 	if err != nil {
 		return err
 	}
 
-	usage, err := fs.DiskUsage(ctx, o.upperPath(id))
+	usage, err := fs.DiskUsage(ctx, o.upperPath(id, i.Labels))
 	if err != nil {
 		return err
 	}
@@ -313,30 +375,37 @@ func (o *snapshotter) Remove(ctx context.Context, key string) (err error) {
 		}
 	}()
 
+	_, info, _, err := storage.GetInfo(ctx, key)
+	if err != nil {
+		return fmt.Errorf("failed to get snapshot info: %w", err)
+	}
 	_, _, err = storage.Remove(ctx, key)
 	if err != nil {
 		return fmt.Errorf("failed to remove: %w", err)
 	}
 
 	if !o.asyncRemove {
-		var removals []string
-		removals, err = o.getCleanupDirectories(ctx, t)
-		if err != nil {
-			return fmt.Errorf("unable to get directories for removal: %w", err)
-		}
+		//不删除upperdir
+		if val, found := info.Labels[notebookLabelKey]; !found || val != notebookLabelValue {
+			var removals []string
+			removals, err = o.getCleanupDirectories(ctx, t)
+			if err != nil {
+				return fmt.Errorf("unable to get directories for removal: %w", err)
+			}
 
-		// Remove directories after the transaction is closed, failures must not
-		// return error since the transaction is committed with the removal
-		// key no longer available.
-		defer func() {
-			if err == nil {
-				for _, dir := range removals {
-					if err := os.RemoveAll(dir); err != nil {
-						log.G(ctx).WithError(err).WithField("path", dir).Warn("failed to remove directory")
+			// Remove directories after the transaction is closed, failures must not
+			// return error since the transaction is committed with the removal
+			// key no longer available.
+			defer func() {
+				if err == nil {
+					for _, dir := range removals {
+						if err := os.RemoveAll(dir); err != nil {
+							log.G(ctx).WithError(err).WithField("path", dir).Warn("failed to remove directory")
+						}
 					}
 				}
-			}
-		}()
+			}()
+		}
 
 	}
 
@@ -352,14 +421,14 @@ func (o *snapshotter) Walk(ctx context.Context, fn snapshots.WalkFunc, fs ...str
 	defer t.Rollback()
 	if o.upperdirLabel {
 		return storage.WalkInfo(ctx, func(ctx context.Context, info snapshots.Info) error {
-			id, _, _, err := storage.GetInfo(ctx, info.Name)
+			id, i, _, err := storage.GetInfo(ctx, info.Name)
 			if err != nil {
 				return err
 			}
 			if info.Labels == nil {
 				info.Labels = make(map[string]string)
 			}
-			info.Labels[upperdirKey] = o.upperPath(id)
+			info.Labels[upperdirKey] = o.upperPath(id, i.Labels)
 			return fn(ctx, info)
 		}, fs...)
 	}
@@ -374,6 +443,11 @@ func (o *snapshotter) Cleanup(ctx context.Context) error {
 	}
 
 	for _, dir := range cleanup {
+		if len(o.upperdirRoot) != 0 {
+			if strings.Contains(dir, o.upperdirRoot) {
+				continue
+			}
+		}
 		if err := os.RemoveAll(dir); err != nil {
 			log.G(ctx).WithError(err).WithField("path", dir).Warn("failed to remove directory")
 		}
@@ -418,6 +492,7 @@ func (o *snapshotter) getCleanupDirectories(ctx context.Context, t storage.Trans
 			continue
 		}
 
+		//孤儿目录如果是notebook应该保留
 		cleanup = append(cleanup, filepath.Join(snapshotDir, d))
 	}
 
@@ -469,8 +544,39 @@ func (o *snapshotter) createSnapshot(ctx context.Context, kind snapshots.Kind, k
 		return nil, fmt.Errorf("failed to create snapshot: %w", err)
 	}
 
+	path = filepath.Join(snapshotDir, s.ID)
+	doRename := true
+	//通用限额和共享存储限额,默认xfs文件系统
+	if val, found := s.Labels[notebookLabelKey]; found && val == notebookLabelValue {
+		//rootfs配置限额,view.Prepare,参数key=====系统盘路径
+		if _, found := s.Labels[notebookNameLabelKey]; !found {
+			return nil, fmt.Errorf("failed to get jobID: %s", key)
+		}
+		if val, found := s.Labels[notebookActionKey]; found && val == notebookStartValue {
+			doRename = false
+			//判断系统盘是否准备就绪
+			if _, found := s.Labels[versionKey]; !found {
+				return nil, fmt.Errorf("failed to get found version: %s", key)
+			}
+			o.timeViewDisk(ctx, s.Labels[notebookNameLabelKey], s.Labels[versionKey], DefaultTimerView)
+			td = filepath.Join(o.upperdirRoot, s.Labels[notebookNameLabelKey])
+			isExist, err := o.dirExists(td)
+			if err != nil {
+				return nil, err
+			}
+
+			if !isExist {
+				return nil, fmt.Errorf("no found upperdir: %s", td)
+			}
+
+		} else {
+			path = filepath.Join(o.upperdirRoot, s.Labels[notebookNameLabelKey])
+		}
+
+	}
+
 	if len(s.ParentIDs) > 0 {
-		st, err := os.Stat(o.upperPath(s.ParentIDs[0]))
+		st, err := os.Stat(o.upperPath(s.ParentIDs[0], nil))
 		if err != nil {
 			return nil, fmt.Errorf("failed to stat parent: %w", err)
 		}
@@ -485,10 +591,13 @@ func (o *snapshotter) createSnapshot(ctx context.Context, kind snapshots.Kind, k
 		}
 	}
 
-	path = filepath.Join(snapshotDir, s.ID)
-	if err = os.Rename(td, path); err != nil {
-		return nil, fmt.Errorf("failed to rename: %w", err)
+	if doRename {
+		if err := os.Rename(td, path); err != nil {
+			log.G(ctx).WithError(err).Warnf("failed to rename file %s", path)
+			//return nil, fmt.Errorf("failed to rename %q -> %q: %w", td, path, err)
+		}
 	}
+
 	td = ""
 
 	rollback = false
@@ -497,6 +606,18 @@ func (o *snapshotter) createSnapshot(ctx context.Context, kind snapshots.Kind, k
 	}
 
 	return o.mounts(s), nil
+}
+
+func (o *snapshotter) dirExists(path string) (bool, error) {
+	info, err := os.Stat(path)
+	if err == nil {
+		return info.IsDir(), nil
+	}
+	if os.IsNotExist(err) {
+		return false, nil
+	}
+
+	return false, err
 }
 
 func (o *snapshotter) prepareDirectory(ctx context.Context, snapshotDir string, kind snapshots.Kind) (string, error) {
@@ -517,6 +638,29 @@ func (o *snapshotter) prepareDirectory(ctx context.Context, snapshotDir string, 
 
 	return td, nil
 }
+func (o *snapshotter) timeViewDisk(ctx context.Context, key, version string, interval time.Duration) error {
+	ticker := time.NewTicker(interval)
+	defer ticker.Stop()
+
+	for {
+		select {
+		case <-ticker.C:
+			viewResp, err := o.diskClient.View(ctx, key)
+			if err != nil {
+				fmt.Printf("disk view failed: %s", err)
+				continue
+			}
+
+			fmt.Printf("job id %s disk info: %v\n", key, viewResp.Disk)
+			if viewResp.Disk.Status == "Completed" && viewResp.Disk.Version == version {
+				return nil
+			}
+
+		case <-ctx.Done():
+			return ctx.Err()
+		}
+	}
+}
 
 func (o *snapshotter) mounts(s storage.Snapshot) []mount.Mount {
 	if len(s.ParentIDs) == 0 {
@@ -529,7 +673,7 @@ func (o *snapshotter) mounts(s storage.Snapshot) []mount.Mount {
 
 		return []mount.Mount{
 			{
-				Source: o.upperPath(s.ID),
+				Source: o.upperPath(s.ID, s.Labels),
 				Type:   "bind",
 				Options: []string{
 					roFlag,
@@ -542,13 +686,14 @@ func (o *snapshotter) mounts(s storage.Snapshot) []mount.Mount {
 	options := o.options
 	if s.Kind == snapshots.KindActive {
 		options = append(options,
-			fmt.Sprintf("workdir=%s", o.workPath(s.ID)),
-			fmt.Sprintf("upperdir=%s", o.upperPath(s.ID)),
+			fmt.Sprintf("workdir=%s", o.workPath(s.ID, s.Labels)),
+			fmt.Sprintf("upperdir=%s", o.upperPath(s.ID, s.Labels)),
 		)
+
 	} else if len(s.ParentIDs) == 1 {
 		return []mount.Mount{
 			{
-				Source: o.upperPath(s.ParentIDs[0]),
+				Source: o.upperPath(s.ParentIDs[0], s.Labels),
 				Type:   "bind",
 				Options: []string{
 					"ro",
@@ -560,7 +705,7 @@ func (o *snapshotter) mounts(s storage.Snapshot) []mount.Mount {
 
 	parentPaths := make([]string, len(s.ParentIDs))
 	for i := range s.ParentIDs {
-		parentPaths[i] = o.upperPath(s.ParentIDs[i])
+		parentPaths[i] = o.upperPath(s.ParentIDs[i], nil)
 	}
 
 	options = append(options, fmt.Sprintf("lowerdir=%s", strings.Join(parentPaths, ":")))
@@ -574,12 +719,32 @@ func (o *snapshotter) mounts(s storage.Snapshot) []mount.Mount {
 
 }
 
-func (o *snapshotter) upperPath(id string) string {
+func (o *snapshotter) upperPath(id string, labels map[string]string) string {
+	if labels == nil {
+		return filepath.Join(o.root, "snapshots", id, "fs")
+	}
+	if _, found := labels[notebookLabelKey]; found {
+		return o.backupUpperPath(labels)
+	}
 	return filepath.Join(o.root, "snapshots", id, "fs")
 }
 
-func (o *snapshotter) workPath(id string) string {
+func (o *snapshotter) workPath(id string, labels map[string]string) string {
+	if labels == nil {
+		return filepath.Join(o.root, "snapshots", id, "work")
+	}
+	if _, found := labels[notebookLabelKey]; found {
+		return o.backupWorkPath(labels)
+	}
 	return filepath.Join(o.root, "snapshots", id, "work")
+}
+
+func (o *snapshotter) backupWorkPath(labels map[string]string) string {
+	return filepath.Join(o.upperdirRoot, labels[notebookNameLabelKey], "work")
+}
+
+func (o *snapshotter) backupUpperPath(labels map[string]string) string {
+	return filepath.Join(o.upperdirRoot, labels[notebookNameLabelKey], "fs")
 }
 
 // Close closes the snapshotter
